+++ src/azure/storage/container/mod.rs
@@ -7,7 +7,7 @@ use azure::core::enumerations;
 use azure::core::parsing::{traverse, cast_must, cast_optional};
 use azure::core::incompletevector::IncompleteVector;
 
-use azure::storage::{LeaseStatus, LeaseState, LeaseDuration};
+use azure::core::lease::{LeaseStatus, LeaseState, LeaseDuration};
 use azure::storage::client::Client;
 
 use hyper::header::Headers;
@@ -116,6 +116,9 @@ impl Container {
 
         Ok(())
     }
+
+    // TODO
+    // pub fn get_acl(c : &Client, gao : &GetAclOptions)
 
     pub fn list(c: &Client,
                 lco: &ListContainerOptions)